+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Fixed
 
-* Improve UX in case when loading of keys and spices fail
+* Improve UX in case when loading of keys and spices fails
+* Improve UX in case when loading of secretmap fails
+* Improve UX in case when loading of treasure from secretmap fails
 
 # 2.0.2 - 3.12.2020.